Change file mode for /etc/nixos to 000
Root can read and write to whatever it wants, whenever it wants, so 600 is technically the same as 000 if it's still owned by root. But this makes it identical to the permissions used by /etc/shadow by default. 

Basically, there is no practical effect from before but it doesn't hurt and it just makes a little more sense.